Refactor the crate so its fully async
The current version no longer creates a dedicated OS thread and also doesn’t have to be mutable to listen for incomming signals anymore.

The internals are now mostly a wrapper around the `CancellationToken` (that can be found in the `tokio-util` crate), but with the addition that it's being triggered (cancelled) by OS signals.

This version also renamed some of the methods, so version v0.3.0 will have breaking changes!